re organize units dir files
