+++ src/latex2sympy/units/dimensions.py
@@ -1,0 +1,5 @@
+from sympy.physics.units import Dimension
+
+# define new dimensions
+solid_angle = Dimension("solid_angle")
+luminous_flux = Dimension("luminous_flux")
