+++ src/latex2sympy/utils/expression.py
@@ -1,6 +1,24 @@
 import sympy
 
 
+def is_or_contains_instance(expr, type):
+    '''
+    Determine if the expression `expr` is or contains an instance of `type`
+    '''
+    if expr is None:  # pragma: no cover
+        return False
+    if isinstance(expr, type):
+        return True
+    if not hasattr(expr, 'args') or len(expr.args) <= 0:  # pragma: no cover
+        return False
+
+    for arg in expr.args:
+        if is_or_contains_instance(arg, type):
+            return True
+
+    return False
+
+
 def create_rational_or_number(text):
     try:
         return sympy.Rational(text)
